Add branch (headRefName) source filter
Allows setting a "branch" in the source configuration, that filters the
PR versions down to only those with a matching headRefName.

When using instanced pipelines (one-pipeline-per-pr) it is desirable to
monitor changes only for an individual PR and not all PRs.

Ideally we could filter by PR-number or some other unique ID, however
the GraphQL API does not expose this, and we want to avoid filtering
client-side.

Note: There _should_ be a one-to-one mapping of headRefName to PR in
most real-world usage, and so for this "one-pipeline-per-pr" use case,
it is probably "good enough". However there is nothing stopping branch
names getting reused over time, so you should not consider setting
"branch" the same as "pinning" a resource version, it is merely a
filter.