Allow cloning the repository via tarball
The default behaviour of this resource is to use `git clone` on "in", cloning the base branch and then switching to the PR branch. This can be incredibly slow on very large repositories (like Monzo's internal monorepo). This applies a fix we've used elsewhere which allows using the Github API to generate a repository tarball link, which we then use to stream files to the filesystem.

This also updates the version of google/go-github as we need to use client.BareDo.